requirements.txt
yfinance
prefect
matplotlib
google-api-python-client
google-cloud
google-cloud-storage